backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
sqlalchemy>=2.0
alembic>=1.13
pydantic-settings>=2.3
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt<4
python-dateutil>=2.9
